runtime: start Qwen3.6 MoE runtime module
